# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fuel Nutrition API",
    "version": "1.0.0",
    "description": "Nutrition tracking API for the Fuel AI coach. Logs food, tracks weight, manages goals, and provides coaching insights."
  },
  "servers": [
    {
      "url": "https://fuel-2j8v.onrender.com/api",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/goal": {
      "get": {
        "tags": [
          "Goal"
        ],
        "summary": "Get current goal",
        "description": "Returns the active user goal configuration (mode, target weight, start weight, start date).",
        "operationId": "getGoal",
        "responses": {
          "200": {
            "description": "Goal configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoal"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Goal"
        ],
        "summary": "Set or update goal",
        "description": "Overwrite the active goal. Missing optional fields fall back to the existing values.",
        "operationId": "setGoal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mode",
                  "targetWeight"
                ],
                "properties": {
                  "mode": {
                    "$ref": "#/components/schemas/GoalMode"
                  },
                  "targetWeight": {
                    "type": "number",
                    "exclusiveMinimum": 0,
                    "description": "Target body weight in kg."
                  },
                  "startWeight": {
                    "type": "number",
                    "exclusiveMinimum": 0,
                    "description": "Starting body weight in kg (optional, falls back to existing)."
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date",
                    "description": "ISO start date (optional, falls back to existing)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Goal updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoal"
                }
              }
            }
          },
          "400": {
            "description": "Invalid mode or targetWeight",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/coach": {
      "get": {
        "tags": [
          "Coach"
        ],
        "summary": "Get coach briefing",
        "description": "Deterministic AI coach briefing with prioritized recommendations, check-in summary, macro targets, and talking points — all grounded in the user's training data.",
        "operationId": "getCoachBriefing",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GoalMode"
            },
            "description": "Goal mode."
          },
          {
            "name": "targetWeight",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 78,
              "minimum": 1,
              "maximum": 500
            },
            "description": "Target weight in kg."
          },
          {
            "name": "currentTarget",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 20000
            },
            "description": "Current daily calorie target. Defaults to engine recommendation."
          }
        ],
        "responses": {
          "200": {
            "description": "Coach briefing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachData"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/coach/chat": {
      "post": {
        "tags": [
          "Coach"
        ],
        "summary": "Chat with AI coach (SSE stream)",
        "description": "Server-Sent Events streaming chat with the Gemini-powered AI coach. Can autonomously search and log foods via the `log_food` tool. Requires GEMINI_API_KEY.\n\nEvents are NDJSON (`data: {...}\\n\\n`): thinking, searching, found, logging, logged, reply, done, error.",
        "operationId": "chatWithCoach",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string",
                    "minLength": 1,
                    "description": "User's message to the coach."
                  },
                  "sessionHistory": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ChatMessage"
                    },
                    "description": "Prior conversation context for multi-turn chat."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE event stream. Each line is `data: {\\\"type\\\":\\\"thinking\\\"|\\\"searching\\\"|\\\"found\\\"|\\\"logging\\\"|\\\"logged\\\"|\\\"reply\\\"|\\\"done\\\"|\\\"error\\\"}\\n\\n`."
          },
          "400": {
            "description": "Missing or empty message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Coach unavailable (GEMINI_API_KEY not set)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/coach/chat/sync": {
      "post": {
        "tags": [
          "Coach"
        ],
        "summary": "Chat with AI coach (sync JSON)",
        "description": "Synchronous chat with the Gemini-powered AI coach. Returns plain JSON — use this from ChatGPT Actions (SSE not supported by ChatGPT). Same backend as /coach/chat. Requires GEMINI_API_KEY.",
        "operationId": "chatWithCoachSync",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message"],
                "properties": {
                  "message": {
                    "type": "string",
                    "minLength": 1,
                    "description": "User's message to the coach."
                  },
                  "sessionHistory": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ChatMessage"
                    },
                    "description": "Prior conversation turns for multi-turn chat."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coach reply",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reply": {
                      "type": "string",
                      "description": "Coach's text response."
                    }
                  },
                  "required": ["reply"]
                }
              }
            }
          },
          "400": {
            "description": "Missing or empty message",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "503": {
            "description": "Coach unavailable (GEMINI_API_KEY not set)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/export": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Export all data as JSON or CSV",
        "description": "Downloads all user data as JSON or CSV file.",
        "operationId": "exportData",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ],
              "default": "json"
            },
            "description": "Export format."
          }
        ],
        "responses": {
          "200": {
            "description": "Data file",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exportedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "demoDate": {
                      "type": "string",
                      "format": "date"
                    },
                    "goal": {
                      "$ref": "#/components/schemas/UserGoal"
                    },
                    "weight": {
                      "$ref": "#/components/schemas/WeightData"
                    },
                    "food": {
                      "$ref": "#/components/schemas/FoodDay"
                    }
                  }
                }
              },
              "text/csv": {}
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message."
          }
        }
      },
...
      "GoalMode": {
        "type": "string",
        "enum": [
          "fat-loss",
          "maintenance",
          "lean-bulk",
          "recomp"
        ],
        "description": "Nutrition goal mode."
      },
      "MealSlot": {
        "type": "string",
        "enum": [
          "breakfast",
          "lunch",
          "dinner",
          "snack"
        ],
        "description": "Meal slot for grouping food entries."
      },
...
      "Macros": {
        "type": "object",
        "required": [
          "calories",
          "protein",
          "carbs",
          "fat"
        ],
        "properties": {
          "calories": {
            "type": "number",
            "description": "Energy in kcal."
          },
          "protein": {
            "type": "number",
            "description": "Protein in grams."
          },
          "carbs": {
            "type": "number",
            "description": "Carbohydrates in grams."
          },
          "fat": {
            "type": "number",
            "description": "Fat in grams."
          }
        }
      },
...
      "LoggedFood": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Macros"
          },
          {
            "type": "object",
            "required": [
              "id",
              "date",
              "slot",
              "foodId",
              "name",
              "quantity",
              "servingUnit",
              "loggedAt"
            ],
            "properties": {
              "id": {
                "type": "string",
                "description": "Unique log entry ID (e.g. 'log-1')."
              },
              "date": {
                "type": "string",
                "format": "date",
                "description": "Date logged."
              },
              "slot": {
                "$ref": "#/components/schemas/MealSlot"
              },
              "foodId": {
                "type": "string",
                "description": "Reference to the FoodItem."
              },
              "name": {
                "type": "string"
              },
              "brand": {
                "type": "string"
              },
              "quantity": {
                "type": "number",
                "description": "Number of servings."
              },
              "servingUnit": {
                "type": "string"
              },
              "loggedAt": {
                "type": "string",
                "format": "date-time",
                "description": "ISO timestamp of when it was logged."
              }
            }
          }
        ]
      },
      "MealGroup": {
        "type": "object",
        "required": [
          "slot",
          "label",
          "entries",
          "totals"
        ],
        "properties": {
          "slot": {
            "$ref": "#/components/schemas/MealSlot"
          },
          "label": {
            "type": "string",
            "description": "Human-readable label (e.g. 'Breakfast', 'Lunch')."
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoggedFood"
            }
          },
          "totals": {
            "$ref": "#/components/schemas/Macros",
            "description": "Rolled-up macros for this meal slot."
          }
        }
      },
      "FoodDay": {
        "type": "object",
        "required": [
          "date",
          "target",
          "consumed",
          "remaining",
          "groups"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "target": {
            "$ref": "#/components/schemas/Macros",
            "description": "Daily macro targets."
          },
          "consumed": {
            "$ref": "#/components/schemas/Macros",
            "description": "Sum of all logged entries."
          },
          "remaining": {
            "$ref": "#/components/schemas/Macros",
            "description": "Target minus consumed."
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealGroup"
            },
            "description": "Four meal slot groups (breakfast, lunch, dinner, snack)."
          }
        }
      },
...
      "WeightPoint": {
        "type": "object",
        "required": [
          "date",
          "scale",
          "trend"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "scale": {
            "type": "number",
            "nullable": true,
            "description": "Raw scale weight (null if no weigh-in that day)."
          },
          "trend": {
            "type": "number",
            "description": "Smoothed trend weight."
          }
        }
      },
...
      "WeightStats": {
        "type": "object",
        "required": [
          "currentTrend",
          "latestScale",
          "weeklyRate",
          "totalChange",
          "lowestTrend",
          "highestTrend",
          "entriesLogged",
          "windowDays"
        ],
        "properties": {
          "currentTrend": {
            "type": "number",
            "description": "Latest smoothed trend weight (kg)."
          },
          "latestScale": {
            "type": "number",
            "nullable": true,
            "description": "Most recent raw scale weight (kg)."
          },
          "weeklyRate": {
            "type": "number",
            "description": "Trend change over last 7 days (kg, signed)."
          },
          "totalChange": {
            "type": "number",
            "description": "Trend change across full window (kg, signed)."
          },
          "lowestTrend": {
            "type": "number",
            "description": "Lowest trend weight in window (kg)."
          },
          "highestTrend": {
            "type": "number",
            "description": "Highest trend weight in window (kg)."
          },
          "entriesLogged": {
            "type": "integer",
            "description": "Days with a raw weigh-in in window."
          },
          "windowDays": {
            "type": "integer",
            "description": "Total days in the window."
          }
        }
      },
      "WeightData": {
        "type": "object",
        "required": [
          "date",
          "range",
          "series",
          "stats",
          "unit"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "description": "Date of the newest data point."
          },
          "range": {
            "type": "integer",
            "enum": [
              30,
              60,
              90,
              180
            ],
            "description": "Lookback window."
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeightPoint"
            }
          },
          "stats": {
            "$ref": "#/components/schemas/WeightStats"
          },
          "unit": {
            "type": "string",
            "enum": [
              "kg",
              "lb"
            ],
            "description": "Weight unit."
          }
        }
      },
...
      "UserGoal": {
        "type": "object",
        "required": [
          "mode",
          "targetWeight",
          "startWeight",
          "startDate"
        ],
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/GoalMode"
          },
          "targetWeight": {
            "type": "number",
            "description": "Goal body weight in kg."
          },
          "startWeight": {
            "type": "number",
            "description": "Starting body weight in kg."
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "Goal start date."
          }
        }
      },
      "CoachCheckIn": {
        "type": "object",
        "required": [
          "periodLabel",
          "avgIntake",
          "avgExpenditure",
          "energyBalance",
          "weightTrendDelta",
          "adherence",
          "verdict"
        ],
        "properties": {
          "periodLabel": {
            "type": "string",
            "description": "e.g. 'Jun 18 – Jul 2'."
          },
          "avgIntake": {
            "type": "number",
            "nullable": true,
            "description": "Mean logged intake (kcal)."
          },
          "avgExpenditure": {
            "type": "number",
            "description": "Mean expenditure (kcal)."
          },
          "energyBalance": {
            "type": "number",
            "nullable": true,
            "description": "Intake minus expenditure (kcal)."
          },
          "weightTrendDelta": {
            "type": "number",
            "description": "Trend change across window (kg, signed)."
          },
          "adherence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "Fraction of days on target."
          },
          "verdict": {
            "type": "string",
            "description": "Narrative assessment."
          }
        }
      },
      "CoachTargets": {
        "type": "object",
        "required": [
          "current",
          "recommended",
          "delta",
          "protein",
          "carbs",
          "fat",
          "rationale"
        ],
        "properties": {
          "current": {
            "type": "number",
            "description": "Current presumed target (kcal)."
          },
          "recommended": {
            "type": "number",
            "description": "Engine recommendation (kcal)."
          },
          "delta": {
            "type": "number",
            "description": "recommended - current (signed kcal)."
          },
          "protein": {
            "type": "number",
            "description": "Recommended protein (g)."
          },
          "carbs": {
            "type": "number",
            "description": "Recommended carbs (g)."
          },
          "fat": {
            "type": "number",
            "description": "Recommended fat (g)."
          },
          "rationale": {
            "type": "string"
          }
        }
      },
      "CoachRecommendation": {
        "type": "object",
        "required": [
          "id",
          "tone",
          "category",
          "title",
          "rationale",
          "action",
          "priority"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "tone": {
            "type": "string",
            "enum": [
              "positive",
              "neutral",
              "warning",
              "action"
            ]
          },
          "category": {
            "type": "string",
            "enum": [
              "calories",
              "macros",
              "activity",
              "adherence",
              "recovery"
            ]
          },
          "title": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "delta": {
            "type": "object",
            "required": [
              "label",
              "value",
              "unit"
            ],
            "properties": {
              "label": {
                "type": "string"
              },
              "value": {
                "type": "number"
              },
              "unit": {
                "type": "string"
              }
            }
          },
          "priority": {
            "type": "integer",
            "description": "1 = highest priority."
          }
        }
      },
      "CoachData": {
        "type": "object",
        "required": [
          "generatedAt",
          "mode",
          "headline",
          "summary",
          "focus",
          "confidence",
          "checkIn",
          "targets",
          "recommendations",
          "talkingPoints"
        ],
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "mode": {
            "$ref": "#/components/schemas/GoalMode"
          },
          "headline": {
            "type": "string",
            "description": "One-line greeting."
          },
          "summary": {
            "type": "string",
            "description": "2–3 sentence assessment."
          },
          "focus": {
            "type": "string",
            "description": "Single biggest lever."
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "checkIn": {
            "$ref": "#/components/schemas/CoachCheckIn"
          },
          "targets": {
            "$ref": "#/components/schemas/CoachTargets"
          },
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CoachRecommendation"
            },
            "description": "Sorted by priority."
          },
          "talkingPoints": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Reflective prompts (max 4)."
          }
        }
      },
      "ChatMessage": {
        "type": "object",
        "required": [
          "role",
          "text"
        ],
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "model"
            ],
            "description": "Who sent the message."
          },
          "text": {
            "type": "string",
            "description": "Message content."
          }
        }
      }
    }
  }
}