# backend/src/routes/coach.ts
import { Router } from 'express';
import type { GoalMode } from '@nutrition/types';
import { buildCoach } from '../domain/coach';
import { buildDailyRecords } from '../domain/dailyRecords';
import { isCoachEnabled, chatWithCoach } from '../domain/geminiCoach';
import { isTrainingDay } from '../domain/trainingDay';
import { DEMO_ANCHOR_DATE } from '../domain/sampleData';

const router = Router();

const VALID_MODES: GoalMode[] = ['fat-loss', 'maintenance', 'lean-bulk', 'recomp'];

function normalizeMode(value: unknown): GoalMode {
  return typeof value === 'string' && (VALID_MODES as string[]).includes(value)
    ? (value as GoalMode)
    : 'fat-loss';
}

function normalizeTargetWeight(value: unknown): number {
  const n = typeof value === 'string' ? Number(value) : NaN;
  return Number.isFinite(n) && n > 0 && n <= 500 ? n : 78;
}

function normalizeTarget(value: unknown): number | undefined {
  const n = typeof value === 'string' ? Number(value) : NaN;
  return Number.isFinite(n) && n > 0 && n <= 20000 ? n : undefined;
}

router.get('/coach', (req, res) => {
  try {
    const mode = normalizeMode(req.query.mode);
    const targetWeight = normalizeTargetWeight(req.query.targetWeight);
    const currentTarget = normalizeTarget(req.query.currentTarget);
    res.json(
      buildCoach(buildDailyRecords(), { mode, targetWeight, currentTarget, trainingDay: isTrainingDay(DEMO_ANCHOR_DATE) }),
    );
  } catch {
    res.status(500).json({ error: 'Internal server error' });
  }
});

/**
 * POST /api/coach/chat — SSE streaming endpoint.
 * Emits newline-delimited JSON events:
 *   {"type":"thinking"}
 *   {"type":"searching","query":"eggs"}
 *   {"type":"found","name":"Egg","calories":78,...}
 *   {"type":"logging","name":"Egg","slot":"breakfast","quantity":2}
 *   {"type":"logged","name":"Egg","slot":"breakfast","calories":156,...}
 *   {"type":"reply","text":"...coach reply..."}
 *   {"type":"done"}
 *   {"type":"error","message":"..."}
 */
router.post('/coach/chat', async (req, res) => {
  if (!isCoachEnabled()) {
    res.status(503).json({ error: 'Coach unavailable. Set GEMINI_API_KEY.' });
    return;
  }

  const { message, sessionHistory } = req.body ?? {};
  if (typeof message !== 'string' || message.trim().length === 0) {
    res.status(400).json({ error: 'Message is required.' });
    return;
  }

  // SSE headers
  res.setHeader('Content-Type', 'text/event-stream');
  res.setHeader('Cache-Control', 'no-cache');
  res.setHeader('Connection', 'keep-alive');
  res.flushHeaders();

  let aborted = false;
  req.on('close', () => { aborted = true; });

  const send = (event: object) => {
    if (aborted) return;
    res.write(`data: ${JSON.stringify(event)}\n\n`);
  };

  try {
    const userId = req.headers['x-user-id'] as string | undefined;
    await chatWithCoach(message, sessionHistory ?? [], send, userId);
    send({ type: 'done' });
  } catch (e) {
    send({ type: 'error', message: e instanceof Error ? e.message : 'Internal server error' });
  } finally {
    res.end();
  }
});

/**
 * POST /api/coach/chat/sync — synchronous JSON endpoint for ChatGPT Actions.
 * ChatGPT HTTP Actions cannot consume SSE — this returns {reply} as plain JSON.
 * Uses the same Gemini backend; SSE progress events are collected and discarded.
 */
router.post('/coach/chat/sync', async (req, res) => {
  if (!isCoachEnabled()) {
    res.status(503).json({ error: 'Coach unavailable. Set GEMINI_API_KEY.' });
    return;
  }

  const { message, sessionHistory } = req.body ?? {};
  if (typeof message !== 'string' || message.trim().length === 0) {
    res.status(400).json({ error: 'Message is required.' });
    return;
  }

  try {
    const userId = req.headers['x-user-id'] as string | undefined;
    const reply = await chatWithCoach(message, sessionHistory ?? [], undefined, userId);
    res.json({ reply });
  } catch (e) {
    res.status(500).json({ error: e instanceof Error ? e.message : 'Internal server error' });
  }
});

export default router;